prueba-u-net-2/requirements.txt
roboflow==1.1.31
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
albumentations>=1.3.0
tqdm>=4.65.0
PyYAML>=6.0 